Fixes for lint & cleanup
